requirements.txt
ipykernel==6.28.0
ipython==8.19.0
jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.6.1
matplotlib==3.8.2
numpy==1.26.2
optuna==3.5.0
pandas==2.1.4
pip==23.3.2
pytorch==2.1.2
pytorch-lightning==2.1.3
scanpy==1.10.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
sqlalchemy==2.0.27
statsmodels==0.14.1
umap-learn==0.5.5
wheel==0.42.0
